Update Prometheus metrics + Add Grafana sample (#339)
* Add time metrics

* Fix prometheus server parsing errors + fix cmd arg

* cmd arg here :)

* increase alarm interval

* fix naming

* add sample grafana dashboard

* add grafana info

* fix format issues

* add instance

+++ src/utils/BaseServerStats.cpp
@@ -40,7 +40,7 @@ void BaseServerStats::initBaseStats(const std::shared_ptr<prometheus::Registry> 
 							  .Add({});
 
 	// Set defaults
-	_minProcessingTime->Set(std::numeric_limits<double>::max());
+	_minProcessingTime->Set(std::numeric_limits<int>::max());
 }
 
 void BaseServerStats::consumeBaseStats(uint64_t succeeded, uint64_t failed, double processingTime)
